Re-factor to fix trunk usage
Neutron trunk's is not a trunk network. A trunk is a special port
with subports. The subports should be on separate networks. When trunk
parent and subports are ports on the same network there is no actual
isolation. Effectively VLAN tags are replaced with the networks
segmentation ID (vlan/vxlan) - i.e if parent and subports are on the
same network it's all the same broadcast domain.

With this change a network and subnet is created for all networks, also
networks that are only attached to instances using a trunk port.

Additionally his change:
* Enables the possibility to attach additional
  ports to instances - either as additional trunk port or
  regular port.
* Adds experimental IPv6 single/dual stack support in some places.
  (This needs further work ...)

Changes required in the input model (network-mapper):
* To enable multiple port attachments and the possibility for multiple
  ports to be a trunk parent the values `is_trunk_parent` (boolean) and
  `trunk_parent` (string) has been added in the networks schema for
  instances.

   Example:
    ---
    instances:
      compute-0:
        networks:
          default:
            ip_v4: 192.168.122.11
            is_trunk_parent: true  <--
          storage:
            ip_v4: 172.18.0.11
            vlan_id: 21
            trunk_parent: default   <--

NOTE! The code to create a router has been remvoed in this commit.
      A follow up commit will add routers.

Resolves: https://issues.redhat.com/browse/OSPRH-5089