Run in_zuul tasks nested on the controller
We have issues with ansible (Openstack.Cloud) collection version
available in the zuul environment.

To enable use of the Openstack.Cloud v2.x ansible and the collection
is installed in a virtualenv on the controller node. The tasks using
openstack.cloud modules run using a nested ansiple playbook executing
on the controller node.