Install openstacksdk in users local env
Something is not quite right with the openstack.cloud collection,
it does not look in the virtualenv for openstacksdk. Workaround
this by pip installing openstacksdk in the users.

It seems in CI something else is installing the openstacksdk, so
the task to install it in users local env is conditioned on the
'local' or 'inject' tag being used.