Include skip_nm ifaces in crc_ci_bootstrap_networks_out
With PR: https://github.com/openstack-k8s-operators/ci-framework/pull/1218
The testproject now fails on:

TASK [Ensure we have needed bits for compute when needed that=[
   "'compute-0' in crc_ci_bootstrap_networks_out",
  "'default' in crc_ci_bootstrap_networks_out['compute-0']"]] ***

fatal: [localhost]: FAILED! => {
    "assertion": "'compute-0' in crc_ci_bootstrap_networks_out",
    "changed": false,
    "evaluated_to": false,
    "msg": "Assertion failed"
}

I had not realized that there was an expectation that skip_nm interfaces
where included in the output.

This PR changes this behaviour, so that skip_nm nodes/interfaces are
included in crc_ci_bootstrap_networks_out.