Add missing vars to in_zuul-vars file
When calling the network clenaup, the nested call to the
in_zuul_cleanup-playbook.yaml playbook fails because
cifmw_bootstrap_create_application_credential is not  defined, even
though it has a default in the role. This change adds it to the
in_zuul-vars.yaml file so that it can be exposed to the nested playbook.